Work around JSONB/Bson deserialization issue

+++ src/laniakea/db/schema/spears.d
@@ -70,8 +70,8 @@ struct SpearsOldBinaries {
  * Age requirements for a package to migrate
  **/
 struct SpearsAgePolicy {
-    uint currentAge;
-    uint requiredAge;
+    long currentAge;
+    long requiredAge;
 }
 
 /**
